+++ test/azure/low-level/Expected/AcceptanceTests/AzureParameterGroupingLowLevel/azureparametergroupinglowlevel/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._auto_rest_parameter_grouping_test_service import AutoRestParameterGroupingTestService
+from ._client import AutoRestParameterGroupingTestService
 
 try:
     from ._patch import __all__ as _patch_all