+++ test/azure/low-level/Expected/AcceptanceTests/HeadLowLevel/headlowlevel/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._auto_rest_head_test_service import AutoRestHeadTestService
+from ._client import AutoRestHeadTestService
 
 try:
     from ._patch import __all__ as _patch_all