+++ test/vanilla/legacy/Expected/AcceptanceTests/SecurityAadSwagger/securityaadswagger/aio/_autorest_security_aad.py
@@ -7,44 +7,50 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, Awaitable, Optional, TYPE_CHECKING
+from typing import Any, Awaitable, TYPE_CHECKING
+
+from msrest import Deserializer, Serializer
 
 from azure.core import AsyncPipelineClient
 from azure.core.rest import AsyncHttpResponse, HttpRequest
-from msrest import Deserializer, Serializer
 
-from ._configuration import ObjectTypeClientConfiguration
+from ._configuration import AutorestSecurityAadConfiguration
+from .operations import AutorestSecurityAadOperationsMixin
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from typing import Dict
 
+    from azure.core.credentials_async import AsyncTokenCredential
 
-class ObjectTypeClient:
-    """Service client for testing basic type: object swaggers.
 
-    :keyword endpoint: Service URL. Default value is 'http://localhost:3000'.
-    :paramtype endpoint: str
+class AutorestSecurityAad(AutorestSecurityAadOperationsMixin):
+    """Autorest Security Aad REST APIs.
+
+    :param credential: Credential needed for the client to connect to Azure.
+    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
+    :param base_url: Service URL. Default value is "http://localhost:3000".
+    :type base_url: str
     """
 
-    def __init__(self, *, endpoint: str = "http://localhost:3000", **kwargs: Any) -> None:
-        self._config = ObjectTypeClientConfiguration(**kwargs)
-        self._client = AsyncPipelineClient(base_url=endpoint, config=self._config, **kwargs)
+    def __init__(
+        self, credential: "AsyncTokenCredential", base_url: str = "http://localhost:3000", **kwargs: Any
+    ) -> None:
+        self._config = AutorestSecurityAadConfiguration(credential=credential, **kwargs)
+        self._client = AsyncPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
-        self._serialize = Serializer()
-        self._deserialize = Deserializer()
+        client_models = {}  # type: Dict[str, Any]
+        self._serialize = Serializer(client_models)
+        self._deserialize = Deserializer(client_models)
         self._serialize.client_side_validation = False
 
-    def send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
+    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
-        We have helper methods to create requests specific to this service in `mergepatchjsonlowlevel.rest`.
-        Use these helper methods to create the request you pass to this method.
-
-        >>> from mergepatchjsonlowlevel.rest import build_patch_single_request
-        >>> request = build_patch_single_request(json=json, content=content, **kwargs)
-        <HttpRequest [PATCH], url: '/mergePatchJson/single'>
-        >>> response = await client.send_request(request)
+        >>> from azure.core.rest import HttpRequest
+        >>> request = HttpRequest("GET", "https://www.example.org/")
+        <HttpRequest [GET], url: 'https://www.example.org/'>
+        >>> response = await client._send_request(request)
         <AsyncHttpResponse: 200 OK>
 
         For more information on this code flow, see https://aka.ms/azsdk/python/protocol/quickstart
@@ -63,7 +69,7 @@ class ObjectTypeClient:
     async def close(self) -> None:
         await self._client.close()
 
-    async def __aenter__(self) -> "ObjectTypeClient":
+    async def __aenter__(self) -> "AutorestSecurityAad":
         await self._client.__aenter__()
         return self