+++ test/vanilla/low-level/Expected/AcceptanceTests/BodyBinaryLowLevel/bodybinarylowlevel/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._binarywithcontent_typeapplicationjson import BinaryWithContentTypeApplicationJson
+from ._client import BinaryWithContentTypeApplicationJson
 
 try:
     from ._patch import __all__ as _patch_all