+++ test/vanilla/low-level/Expected/AcceptanceTests/BodyFormUrlEncodedDataLowLevel/bodyformurlencodeddatalowlevel/aio/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._body_forms_data_url_encoded import BodyFormsDataURLEncoded
+from ._client import BodyFormsDataURLEncoded
 
 try:
     from ._patch import __all__ as _patch_all