+++ test/vanilla/low-level/Expected/AcceptanceTests/MediaTypesLowLevel/mediatypeslowlevel/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._media_types_client import MediaTypesClient
+from ._client import MediaTypesClient
 from ._version import VERSION
 
 __version__ = VERSION