+++ test/vanilla/low-level/Expected/AcceptanceTests/ParameterizedEndpointLowLevel/parameterizedendpointlowlevel/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._parmaterized_endpoint_client import ParmaterizedEndpointClient
+from ._client import ParmaterizedEndpointClient
 from ._version import VERSION
 
 __version__ = VERSION